checkpointer: change activeGracePeriod logic.
This changes activeGracePeriod to move to inactiveGracePeriod when the
apiserver is up. This is to allow the checkpointer to get out of the
way of the real apiserver when it comes back, while still protecting
against that real apiserver going back down shortly after.

Before this change, a checkpoint could remain active for the entire
grace period unnecessarily if it was replaced by a new pod with a
different name that was successfully running.

After this change, a checkpoint will only get cleaned up after the
apiserver has been up and stable for an entire grace period's length.